Tentative fix for issue with lftools 0.20.0 and jjb.
Observing the following error
11:33:56   File "/home/jenkins/.local/lib/python2.7/site-packages/lftools/deploy.py", line 166, in copy_archives
11:33:56     for file_or_dir in os.listdir(archives_dir):
11:33:56 OSError: [Errno 2] No such file or directory: '/w/workspace/music-mdbc-master-verify-java/archives'

Adding a check if dir/file exists before executing for loop.

+++ lftools/deploy.py
@@ -163,13 +163,17 @@ def copy_archives(workspace, pattern=None):
 
     log.debug('Copying files from {} with pattern \'{}\' to {}.'.format(
         workspace, pattern, dest_dir))
-    for file_or_dir in os.listdir(archives_dir):
-        f = os.path.join(archives_dir, file_or_dir)
-        try:
-            log.debug('Moving {}'.format(f))
-            shutil.move(f, dest_dir)
-        except shutil.Error as e:
-            log.warn(e)
+
+    if not os.path.is_dir(archives_dir) and not os.path.is_file(archives_dir) :
+        log.debug('Archives dir/file {} don't exist.'.format(archives_dir))
+	else:
+		for file_or_dir in os.listdir(archives_dir):
+			f = os.path.join(archives_dir, file_or_dir)
+			try:
+				log.debug('Moving {}'.format(f))
+				shutil.move(f, dest_dir)
+			except shutil.Error as e:
+				log.warn(e)
 
     if pattern is None:
         return
